file module done

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const {validationResult} = require('express-validator')
 const Product = require('../models/product');
-
+const fileHelper = require('../util/file');
+const product = require('../models/product');
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
@@ -14,11 +15,26 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
 
+  if(!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: "Attached file dose not type of image",
+      validationErrors: []
+    });
+  }
   if (!errors.isEmpty()) {
     console.log(errors.array());
     return res.status(422).render('admin/edit-product', {
@@ -28,7 +44,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -36,6 +51,7 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+  const imageUrl = image.path;
   const product = new Product({
     title: title,
     price: price,
@@ -87,7 +103,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -100,7 +116,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -117,8 +132,11 @@ exports.postEditProduct = (req, res, next) => {
     }
     product.title = updatedTitle,
     product.price = updatedPrice,
-    product.imageUrl = updatedImageUrl,
     product.description = updatedDesc
+    if(image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     return product.save().then(result => {
       console.log('UPDATED PRODUCT!');
       res.redirect('/admin/products');
@@ -149,8 +167,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(result => {
+  Product.findById(prodId).then(product => {
+    if(!product) {
+      return next(new Error('Product Not Found!'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({_id: prodId, userId: req.user._id});
+  }).then(result => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
